requirements.txt
pandas
PySastrawi
torch
transformers
regex
fastapi
uvicorn
pydantic